Modify main.js
Change the behaviour when the prop "commits" of component changed.
When the commits property changes, the compoment used to change its state
in its render function, which will cause a warning. So the function
"componentWillReceiveProps" is added to detect the property changing and
renew the state before the render function is called. See
http://facebook.github.io/react/docs/component-specs.html#lifecycle-methods
for more information.

	modified:   static/build/main.js
	modified:   static/src/main.js